LayoutTests/streams/reference-implementation/count-queuing-strategy.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script src='resources/streams-utils.js'></script>
<script>
test(function() {
    new CountQueuingStrategy({ highWaterMark: 4 });
}, 'Can construct a CountQueuingStrategy with a valid high water mark');

test(function() {
    for (var highWaterMark of [-Infinity, NaN, 'foo', {}, function () {}]) {
        var strategy = new CountQueuingStrategy({ highWaterMark });
        assert_true(Object.is(strategy.highWaterMark, highWaterMark), highWaterMark + ' gets set correctly');
    }
}, 'Can construct a CountQueuingStrategy with any value as its high water mark');

test(function() {
    var highWaterMark = 1;
    var highWaterMarkObjectGetter = {
        get highWaterMark() { return highWaterMark; },
    };
    var error = new Error('wow!');
    var highWaterMarkObjectGetterThrowing = {
        get highWaterMark() { throw error; },
    };
    assert_throws(new TypeError(), function() { new CountQueuingStrategy(); }, 'construction fails with undefined');
    assert_throws(new TypeError(), function() { new CountQueuingStrategy(null); }, 'construction fails with null');
    new CountQueuingStrategy('potato'); // Construction succeeds with a random non-object type.
    new CountQueuingStrategy({}); // Construction succeeds with an object without a highWaterMark property.
    new CountQueuingStrategy(highWaterMarkObjectGetter); // Construction succeeds with an object with a highWaterMark getter.
    assert_throws(error, function() { new CountQueuingStrategy(highWaterMarkObjectGetterThrowing); },
                  'construction fails with an object with a throwing highWaterMark getter');
}, 'CountQueuingStrategy constructor behaves as expected with wrong arguments');

test(function() {
    var size = 1024;
    var chunk = { byteLength: size };
    var chunkGetter = {
        get byteLength() { return size; },
    }
    var error = new Error('wow!');
    var chunkGetterThrowing = {
        get byteLength() { throw error; },
    }
    assert_equals(CountQueuingStrategy.prototype.size(), 1, 'size returns 1 with undefined');
    assert_equals(CountQueuingStrategy.prototype.size(null), 1, 'size returns 1 with null');
    assert_equals(CountQueuingStrategy.prototype.size('potato'), 1, 'size returns 1 with non-object type');
    assert_equals(CountQueuingStrategy.prototype.size({}), 1, 'size returns 1 with empty object');
    assert_equals(CountQueuingStrategy.prototype.size(chunk), 1, 'size returns 1 with a chunk');
    assert_equals(CountQueuingStrategy.prototype.size(chunkGetter), 1, 'size returns 1 with chunk getter');
    assert_equals(CountQueuingStrategy.prototype.size(chunkGetterThrowing), 1, 'size returns 1 with chunk getter that throws');
}, 'CountQueuingStrategy size behaves as expected with wrong arguments');

test(function() {
    var strategy = new CountQueuingStrategy({ highWaterMark: 4 });

    assert_object_equals(Object.getOwnPropertyDescriptor(strategy, 'highWaterMark'),
                         { value: 4, writable: true, enumerable: true, configurable: true },
                         'highWaterMark property should be a data property with the value passed the connstructor');
    assert_equals(typeof strategy.size, 'function');
}, 'CountQueuingStrategy instances have the correct properties');

test(function() {
    new ReadableStream({}, new CountQueuingStrategy({ highWaterMark: 4 }));
}, 'Can construct a readable stream with a valid CountQueuingStrategy');

var test1 = async_test('Correctly governs the return value of a ReadableStream\'s enqueue function (HWM = 0)');
test1.step(function() {
    var controller;
    var rs = new ReadableStream({
        start: function(c) {
            controller = c;
        }
    },
    new CountQueuingStrategy({ highWaterMark: 0 })
    );
    var reader = rs.getReader();

    assert_equals(controller.desiredSize, 0, '0 reads, 0 enqueues: desiredSize should be 0');
    controller.enqueue('a');
    assert_equals(controller.desiredSize, -1, '0 reads, 1 enqueue: desiredSize should be -1');
    controller.enqueue('b');
    assert_equals(controller.desiredSize, -2, '0 reads, 2 enqueues: desiredSize should be -2');
    controller.enqueue('c');
    assert_equals(controller.desiredSize, -3, '0 reads, 3 enqueues: desiredSize should be -3');
    controller.enqueue('d');
    assert_equals(controller.desiredSize, -4, '0 reads, 4 enqueues: desiredSize should be -4');

    reader.read().then(test1.step_func(function(result) {
        assert_object_equals(result, { value: 'a', done: false }, '1st read gives back the 1st chunk enqueued (queue now contains 3 chunks)');
        return reader.read();
    })).then(test1.step_func(function(result) {
        assert_object_equals(result, { value: 'b', done: false }, '2nd read gives back the 2nd chunk enqueued (queue now contains 2 chunks)');
        return reader.read();
    })).then(test1.step_func(function(result) {
        assert_object_equals(result, { value: 'c', done: false }, '3rd read gives back the 3rd chunk enqueued (queue now contains 1 chunk)');

        assert_equals(controller.desiredSize, -1, '3 reads, 4 enqueues: desiredSize should be -1');
        controller.enqueue('e');
        assert_equals(controller.desiredSize, -2, '3 reads, 5 enqueues: desiredSize should be -2');

        return reader.read();
    })).then(test1.step_func(function(result) {
        assert_object_equals(result, { value: 'd', done: false }, '4th read gives back the 4th chunk enqueued (queue now contains 1 chunks)');
        return reader.read();
    })).then(test1.step_func(function(result) {
        assert_object_equals(result, { value: 'e', done: false }, '5th read gives back the 5th chunk enqueued (queue now contains 0 chunks)');

        assert_equals(controller.desiredSize, 0, '5 reads, 5 enqueues: desiredSize should be 0');
        controller.enqueue('f');
        assert_equals(controller.desiredSize, -1, '5 reads, 6 enqueues: desiredSize should be -1');
        controller.enqueue('g');
        assert_equals(controller.desiredSize, -2, '5 reads, 7 enqueues: desiredSize should be -2');

        test1.done();
    })).catch(test1.step_func(function(e) { assert_unreached(e); } ));
});

var test2 = async_test('Correctly governs a ReadableStreamController\'s desiredSize property (HWM = 1)');
test2.step(function() {
    var controller;
    var rs = new ReadableStream({
        start: function(c) {
            controller = c;
        },
    },
    new CountQueuingStrategy({ highWaterMark: 1 })
    );
    var reader = rs.getReader();

    assert_equals(controller.desiredSize, 1, '0 reads, 0 enqueues: desiredSize should be 1');
    controller.enqueue('a');
    assert_equals(controller.desiredSize, 0, '0 reads, 1 enqueue: desiredSize should be 0');
    controller.enqueue('b');
    assert_equals(controller.desiredSize, -1, '0 reads, 2 enqueues: desiredSize should be -1');
    controller.enqueue('c');
    assert_equals(controller.desiredSize, -2, '0 reads, 3 enqueues: desiredSize should be -2');
    controller.enqueue('d');
    assert_equals(controller.desiredSize, -3, '0 reads, 4 enqueues: desiredSize should be -3');

    reader.read().then(test2.step_func(function(result) {
        assert_object_equals(result, { value: 'a', done: false }, '1st read gives back the 1st chunk enqueued (queue now contains 3 chunks)');
        return reader.read();
    })).then(test2.step_func(function(result) {
        assert_object_equals(result, { value: 'b', done: false }, '2nd read gives back the 2nd chunk enqueued (queue now contains 2 chunks)');
        return reader.read();
    })).then(test2.step_func(function(result) {
        assert_object_equals(result, { value: 'c', done: false }, '3rd read gives back the 3rd chunk enqueued (queue now contains 1 chunk)');

        assert_equals(controller.desiredSize, 0, '3 reads, 4 enqueues: desiredSize should be 0');
        controller.enqueue('e');
        assert_equals(controller.desiredSize, -1, '3 reads, 5 enqueues: desiredSize should be -1');

        return reader.read();
    })).then(test2.step_func(function(result) {
        assert_object_equals(result, { value: 'd', done: false }, '4th read gives back the 4th chunk enqueued (queue now contains 1 chunks)');
        return reader.read();
    })).then(test2.step_func(function(result) {
        assert_object_equals(result, { value: 'e', done: false }, '5th read gives back the 5th chunk enqueued (queue now contains 0 chunks)');

        assert_equals(controller.desiredSize, 1, '5 reads, 5 enqueues: desiredSize should be 1');
        controller.enqueue('f');
        assert_equals(controller.desiredSize, 0, '5 reads, 6 enqueues: desiredSize should be 0');
        controller.enqueue('g');
        assert_equals(controller.desiredSize, -1, '5 reads, 7 enqueues: desiredSize should be -1');

        test2.done();
    })).catch(test2.step_func(function(e) { assert_unreached(e); } ));
});

var test3 = async_test('Correctly governs a ReadableStreamController\'s desiredSize property (HWM = 4)');
test3.step(function() {
    var controller;
    var rs = new ReadableStream({
        start: function(c) {
            controller = c;
        }
    },
    new CountQueuingStrategy({ highWaterMark: 4 })
    );
    var reader = rs.getReader();

    assert_equals(controller.desiredSize, 4, '0 reads, 0 enqueues: desiredSize should be 4');
    controller.enqueue('a');
    assert_equals(controller.desiredSize, 3, '0 reads, 1 enqueue: desiredSize should be 3');
    controller.enqueue('b');
    assert_equals(controller.desiredSize, 2, '0 reads, 2 enqueues: desiredSize should be 2');
    controller.enqueue('c');
    assert_equals(controller.desiredSize, 1, '0 reads, 3 enqueues: desiredSize should be 1');
    controller.enqueue('d');
    assert_equals(controller.desiredSize, 0, '0 reads, 4 enqueues: desiredSize should be 0');
    controller.enqueue('e');
    assert_equals(controller.desiredSize, -1, '0 reads, 5 enqueues: desiredSize should be -1');
    controller.enqueue('f');
    assert_equals(controller.desiredSize, -2, '0 reads, 6 enqueues: desiredSize should be -2');


    reader.read().then(test3.step_func(function(result) {
        assert_object_equals(result, { value: 'a', done: false }, '1st read gives back the 1st chunk enqueued (queue now contains 5 chunks)');
        return reader.read();
    })).then(test3.step_func(function(result) {
        assert_object_equals(result, { value: 'b', done: false }, '2nd read gives back the 2nd chunk enqueued (queue now contains 4 chunks)');

        assert_equals(controller.desiredSize, 0, '2 reads, 6 enqueues: desiredSize should be 0');
        controller.enqueue('g');
        assert_equals(controller.desiredSize, -1, '2 reads, 7 enqueues: desiredSize should be -1');

        return reader.read();
    })).then(test3.step_func(function(result) {
        assert_object_equals(result, { value: 'c', done: false }, '3rd read gives back the 3rd chunk enqueued (queue now contains 4 chunks)');
        return reader.read();
    })).then(test3.step_func(function(result) {
        assert_object_equals(result, { value: 'd', done: false }, '4th read gives back the 4th chunk enqueued (queue now contains 3 chunks)');
        return reader.read();
    })).then(test3.step_func(function(result) {
        assert_object_equals(result, { value: 'e', done: false }, '5th read gives back the 5th chunk enqueued (queue now contains 2 chunks)');
        return reader.read();
    })).then(test3.step_func(function(result) {
        assert_object_equals(result, { value: 'f', done: false }, '6th read gives back the 6th chunk enqueued (queue now contains 0 chunks)');

        assert_equals(controller.desiredSize, 3, '6 reads, 7 enqueues: desiredSize should be 3');
        controller.enqueue('h');
        assert_equals(controller.desiredSize, 2, '6 reads, 8 enqueues: desiredSize should be 2');
        controller.enqueue('i');
        assert_equals(controller.desiredSize, 1, '6 reads, 9 enqueues: desiredSize should be 1');
        controller.enqueue('j');
        assert_equals(controller.desiredSize, 0, '6 reads, 10 enqueues: desiredSize should be 0');
        controller.enqueue('k');
        assert_equals(controller.desiredSize, -1, '6 reads, 11 enqueues: desiredSize should be -1');

        test3.done();
    })).catch(test3.step_func(function(e) { assert_unreached(e); } ));
});
</script>
